:sparkles: feat: Adicionando funcao de limpar inputs
