+++ assets/css/cadastro/cadastro.css
@@ -15,3 +15,7 @@ body {
 	display: block;
 	margin: 0.5rem;
 }
+
+.erro {
+	display: none;
+}
